blk-mq: Support for Open-Channel SSDs
Open-channel SSDs are devices which exposes direct access to its
physical flash storage, while maintaining a subset of internal
features in SSDs.

A common SSD consists of a flash translation layer (FTL), bad block
management, and hardware units such as flash controller and host
interface controller and a large amount of flash chips.

LightNVM moves part of the FTL responsibility into the host, allowing
the host to manage data placement, garbage collection and parallelism.

The architecture of LightNVM consists of a core and multiple targets.
The core has the part of the driver that is shared across targets,
initialization/teardown and statistics. Targets are how physical
flash are exposed to the user. This can be as the block device,
key-value store, object-store, or anything else.

Contributions in this patch from:

  Javier Gonzalez <javier@paletta.io>
  Jesper Madsen <jmad@itu.dk>