block: expose init_request_from_bio
When doing internal commands within the a device driver, we like to form
it as a bio, and use the existing infrastructure for handling bio's.
Exposing init_request_from_bio as blk_init_request_from_bio, it allows
us to initialized the request with a bio that we have initialized.