lightnvm: implement logical to physical device mapping
To enable both the logical and physical address to be communicated to
the device, we extend the structure with a phys_sector, that can be used
be the device driver.

For the nvme driver, we overwrite dword 14 and 15 of the NVMe rw
command. This overwrites the end-to-end protection, but allows us to
directly communicate the physical address in a single command to the
device, without using an metadata buffer.

This fixes the rq->__sector hack that we have had applied beforehand.