lightnvm: fix panic when qemu lreadl2ptbl is zero
The kernel panics when the qemu l2p table haven't been initialized.
This is caused by pba being ~0ULL and passed to raddr[] causing an
out-of-bound access.

Additionally use s->nr_pages instead of elba to detect broken tables.
As it is calcualted from the identification, it is a better
crosscheck.