lightnvm: use request tag for inflight tracking
We currently used percpu_ida to get an index into the inflight entries
that was free. This had a smaller array of inflights requests, leading
to better use of resources. However, it also gave problems with number
of requests in progress was larger than the inflight tracking, leading
to spinning in the inflight path.

Therefore we instead preallocate the memory upfront, and uses rq->tag as
index. This can further be optimized when blk-mq some day gains supports
for shrinking/growing the preallocated request pools.